refactor : CalculateService 클래스 및 테스트 구현
CalculateService 클래스 및 테스트 구현
- 로또 개수 계산
- 수익률 계산
